Add radon library and setup.py command support
