+++ radon.cfg
@@ -1,0 +1,3 @@
+[radon]
+order=SCORE
+show_mi=true
