+++ requirements.txt
@@ -5,6 +5,7 @@ attrs==22.1.0
 black==22.10.0
 bytecode==0.13.0
 click==8.1.3
+colorama==0.4.6
 commonmark==0.9.1
 exceptiongroup==1.0.1
 ffmpeg-python==0.2.0
@@ -16,6 +17,7 @@ isort==5.10.1
 jellyfish==0.9.0
 Jinja2==3.1.2
 lazy-object-proxy==1.8.0
+mando==0.6.4
 MarkupSafe==2.1.1
 mccabe==0.7.0
 MutPy-Pynguin==0.7.1
@@ -34,8 +36,10 @@ pynguin==0.27.0
 pyparsing==3.0.9
 pytest==7.2.0
 PyYAML==6.0
+radon==5.1.0
 rich==12.6.0
 simple-parsing==0.0.20
+six==1.16.0
 termcolor==2.1.0
 tomli==2.0.1
 typing-inspect==0.8.0
